# docs/swagger.yaml
basePath: /api/v1
definitions:
  ginx.ResponseData:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
  ginx.ResponseFail:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
...
  schema.RoleDataScopeReq:
    properties:
      dataScope:
        type: string
      deptIds:
        items:
          type: integer
        type: array
      id:
        description: 角色编码
        type: integer
      updateBy:
        description: 更新者
        type: integer
    required:
    - dataScope
    - id
    type: object
  schema.SysApi:
    properties:
      action:
        description: 请求类型
        type: string
      createBy:
        description: 创建者
        type: integer
      createdAt:
        description: 创建时间
        type: string
      deletedAt:
        description: 删除时间
        type: string
      handle:
        description: handle
        type: string
      id:
        description: 主键编码
        type: integer
      path:
        description: 地址
        type: string
      title:
        description: 标题
        type: string
      type:
        description: 接口类型
        type: string
      updateBy:
        description: 更新者
        type: integer
      updatedAt:
        description: 最后更新时间
        type: string
    type: object
...
  schema.SysMenu:
    properties:
      action:
        type: string
      apis:
        items:
          type: integer
        type: array
      breadcrumb:
        type: integer
      children:
        items:
          $ref: '#/definitions/schema.SysMenu'
        type: array
      component:
        type: string
      createBy:
        description: 创建者
        type: integer
      createdAt:
        description: 创建时间
        type: string
      dataScope:
        type: string
      icon:
        type: string
      is_select:
        type: boolean
      isFrame:
        type: integer
      menuId:
        type: integer
      menuName:
        type: string
      menuType:
        type: string
      noCache:
        type: integer
      params:
        type: string
      parentId:
        type: integer
      path:
        type: string
      paths:
        type: string
      permission:
        type: string
      roleId:
        type: integer
      sort:
        type: integer
      sysApi:
        items:
          $ref: '#/definitions/schema.SysApi'
        type: array
      title:
        type: string
      updateBy:
        description: 更新者
        type: integer
      updatedAt:
        description: 最后更新时间
        type: string
      visible:
        type: integer
    type: object
...
  schema.SysPost:
    properties:
      createBy:
        description: 创建者
        type: integer
      dataScope:
        type: string
      params:
        type: string
      postCode:
        description: 岗位代码
        type: string
      postId:
        description: 岗位编号
        type: integer
      postName:
        description: 岗位名称
        type: string
      remark:
        description: 描述
        type: string
      sort:
        description: 岗位排序
        type: integer
      status:
        description: 状态
        type: integer
      updateBy:
        description: 更新者
        type: integer
    type: object
  schema.SysPostDeleteReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
  schema.SysPostInsertReq:
    properties:
      createBy:
        description: 创建者
        type: integer
      postCode:
        type: string
      postName:
        type: string
      remark:
        type: string
      sort:
        type: integer
      status:
        type: integer
    type: object
  schema.SysPostPageResp:
    properties:
      data:
        items:
          $ref: '#/definitions/schema.SysPost'
        type: array
      total:
        type: integer
    type: object
  schema.SysPostUpdateReq:
    properties:
      id:
        type: integer
      postCode:
        type: string
      postName:
        type: string
      remark:
        type: string
      sort:
        type: integer
      status:
        type: integer
      updateBy:
        description: 更新者
        type: integer
    type: object
  schema.SysRole:
    properties:
      admin:
        type: integer
      createBy:
        description: 创建者
        type: integer
      createdAt:
        description: 创建时间
        type: string
      dataScope:
        type: string
      flag:
        type: string
      menuIds:
        items:
          type: integer
        type: array
      params:
        type: string
      remark:
        description: 备注
        type: string
      roleId:
        description: 角色编码
        type: integer
      roleKey:
        description: 角色代码
        type: string
      roleName:
        description: 角色名称
        type: string
      roleSort:
        description: 角色排序
        type: integer
      status:
        type: integer
      sysMenu:
        items:
          $ref: '#/definitions/schema.SysMenu'
        type: array
      updateBy:
        description: 更新者
        type: integer
      updatedAt:
        description: 最后更新时间
        type: string
    type: object
  schema.SysRoleGetPageResp:
    properties:
      data:
        items:
          $ref: '#/definitions/schema.SysRole'
        type: array
      total:
        type: integer
    type: object
  schema.SysRoleInsertReq:
    properties:
      admin:
        type: integer
      createBy:
        description: 创建者
        type: integer
      dataScope:
        type: string
      flag:
        description: 标记
        type: string
      menuIds:
        items:
          type: integer
        type: array
      remark:
        description: 备注
        type: string
      roleKey:
        description: 角色代码
        type: string
      roleName:
        description: 角色名称
        type: string
      roleSort:
        description: 角色排序
        type: integer
      status:
        description: 状态
        type: integer
    type: object
  schema.SysRoleUpdateReq:
    properties:
      admin:
        type: integer
      dataScope:
        type: string
      flag:
        description: 标记
        type: string
      id:
        description: 角色编码
        type: integer
      menuIds:
        items:
          type: integer
        type: array
      remark:
        description: 备注
        type: string
      roleKey:
        description: 角色代码
        type: string
      roleName:
        description: 角色名称
        type: string
      roleSort:
        description: 角色排序
        type: integer
      status:
        description: 状态
        type: integer
      updateBy:
        description: 更新者
        type: integer
    type: object
...
  schema.UpdateStatusReq:
    properties:
      id:
        description: 角色编码
        type: integer
      status:
        description: 状态
        type: integer
      updateBy:
        description: 更新者
        type: integer
    type: object
...
host: 127.0.0.1:8000
info:
  contact:
    email: conjurer888888@gmail.com
    name: conjurer
    url: https:/github.com/dot123
  description: This is a game management background. you can use the api key `ApiKeyAuth`
    to test the authorization filters.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com
  title: gin-gorm-admin API
  version: "1.0"
paths:
...
  /post:
    delete:
      consumes:
      - application/json
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.SysPostDeleteReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            $ref: '#/definitions/ginx.ResponseData'
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 删除岗位
      tags:
      - 岗位
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: pageNum
        type: integer
      - in: query
        maximum: 100
        name: pageSize
        type: integer
      - description: 编码
        in: query
        name: postCode
        type: string
      - description: id
        in: query
        name: postID
        type: integer
      - description: 名称
        in: query
        name: postName
        type: string
      - description: 备注
        in: query
        name: remark
        type: string
      - description: 排序
        in: query
        name: sort
        type: integer
      - description: 状态
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            allOf:
            - $ref: '#/definitions/ginx.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/schema.SysPostPageResp'
              type: object
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 岗位列表数据
      tags:
      - 岗位
    post:
      consumes:
      - application/json
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.SysPostInsertReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            $ref: '#/definitions/ginx.ResponseData'
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 添加岗位
      tags:
      - 岗位
    put:
      consumes:
      - application/json
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.SysPostUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            $ref: '#/definitions/ginx.ResponseData'
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 修改岗位
      tags:
      - 岗位
  /post/{postId}:
    get:
      consumes:
      - application/json
      parameters:
      - description: 编码
        in: path
        name: postId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            allOf:
            - $ref: '#/definitions/ginx.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/schema.SysPost'
              type: object
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 获取岗位信息
      tags:
      - 岗位
...
  /role:
    get:
      consumes:
      - application/json
      parameters:
      - description: 1不是管理员 2是管理员
        in: query
        name: admin
        type: integer
      - in: query
        name: dataScope
        type: string
      - description: 标记
        in: query
        name: flag
        type: string
      - in: query
        name: pageNum
        type: integer
      - in: query
        maximum: 100
        name: pageSize
        type: integer
      - description: 备注
        in: query
        name: remark
        type: string
      - description: 角色编码
        in: query
        name: roleID
        type: integer
      - description: 角色代码
        in: query
        name: roleKey
        type: string
      - description: 角色名称
        in: query
        name: roleName
        type: string
      - description: 角色排序
        in: query
        name: roleSort
        type: integer
      - description: 状态
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            allOf:
            - $ref: '#/definitions/ginx.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/schema.SysRoleGetPageResp'
              type: object
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 角色列表数据
      tags:
      - 角色/Role
    post:
      consumes:
      - application/json
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.SysRoleInsertReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            $ref: '#/definitions/ginx.ResponseData'
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 创建角色
      tags:
      - 角色/Role
    put:
      consumes:
      - application/json
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.SysRoleUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            $ref: '#/definitions/ginx.ResponseData'
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 修改用户角色
      tags:
      - 角色/Role
  /role-datascope:
    put:
      consumes:
      - application/json
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.RoleDataScopeReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            $ref: '#/definitions/ginx.ResponseData'
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 更新角色数据权限
      tags:
      - 角色/Role
  /role-status:
    put:
      consumes:
      - application/json
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.UpdateStatusReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            $ref: '#/definitions/ginx.ResponseData'
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 修改用户角色状态
      tags:
      - 角色/Role
  /role/{roleId}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: roleId
        in: path
        name: roleId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            $ref: '#/definitions/ginx.ResponseData'
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 删除用户角色
      tags:
      - 角色/Role
    get:
      consumes:
      - application/json
      parameters:
      - description: roleId
        in: path
        name: roleId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功结果
          schema:
            allOf:
            - $ref: '#/definitions/ginx.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/schema.SysRole'
              type: object
        "500":
          description: 失败结果
          schema:
            $ref: '#/definitions/ginx.ResponseFail'
      security:
      - ApiKeyAuth: []
      summary: 获取Role数据
      tags:
      - 角色/Role
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"